feat(basic-tracer): move BasicTracr to a separate package
