+++ packages/opentelemetry-basic-tracer/src/index.ts
@@ -13,3 +13,5 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+export * from './types';
