+++ packages/opentelemetry-core/src/index.ts
@@ -17,6 +17,9 @@
 export * from './common/NoopLogger';
 export * from './context/propagation/BinaryTraceContext';
 export * from './context/propagation/HttpTraceContext';
+export * from './common/NoopLogger';
+export * from './platform';
 export * from './resources/Resource';
 export * from './trace/NoopSpan';
 export * from './trace/NoopTracer';
+export * from './trace/sampler/ProbabilitySampler';
