+++ packages/opentelemetry-types/src/trace/tracer.ts
@@ -21,7 +21,6 @@ import { Sampler } from './Sampler';
 import { ScopeManager } from '@opentelemetry/scope-base';
 import { Span } from './span';
 import { SpanOptions } from './SpanOptions';
-import { Logger } from '../common/Logger';
 
 /**
  * Tracer provides an interface for creating {@link Span}s and propagating
@@ -99,10 +98,6 @@ export interface Tracer {
  * TracerConfig provides an interface for configuring a Tracer.
  */
 export interface TracerConfig {
-  /**
-   * User provided logger.
-   */
-  logger?: Logger;
   /**
    * Attributed that will be applied on every span created by Tracer.
    * Useful to add infrastructure and environment information to your spans.
